+++ actix-web-lab/src/lib.rs
@@ -31,6 +31,7 @@ mod catch_panic;
 #[cfg(feature = "cbor")]
 mod cbor;
 mod clear_site_data;
+mod condition_option;
 mod content_length;
 mod csv;
 mod display_stream;